working through bugs in starting pipeline

+++ Standard DB Queries/reset_db.py
@@ -7,6 +7,7 @@ def main():
 
     delete_current_data(conn, cursor)
     init_tables(conn, cursor)
+    init_teams_table(conn, cursor)
 
 
     # Close connection
@@ -23,7 +24,8 @@ def delete_current_data(conn, cursor):
     # Drop each table
     for table in tables:
         table_name = table[0]
-        conn.execute(f"DROP TABLE IF EXISTS {table_name};")
+        if table_name != 'sqlite_sequence':
+            conn.execute(f"DROP TABLE IF EXISTS {table_name};")
 
     # Commit the changes
     conn.commit()
@@ -34,6 +36,16 @@ initialize tables form create_db.sql script (holds table names and all attribute
 def init_tables(conn, cursor):
     # Read and execute the SQL query from the file
     with open('create_db.sql', 'r') as file:
+        sql_query = file.read()
+        cursor.executescript(sql_query)
+
+    # Commit the changes
+    conn.commit()
+
+
+def init_teams_table(conn, cursor):
+    # Reset teams table after db dropped with init_teams.sql
+    with open('init_teams.sql', 'r') as file:
         sql_query = file.read()
         cursor.executescript(sql_query)
